// src/types/graphql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ABIJson: any;
  AppId: any;
  BlockchainData: any;
  BroadcastId: any;
  ChainId: any;
  ChallengeId: any;
  ContentEncryptionKey: any;
  CreateHandle: any;
  Cursor: any;
  DateTime: any;
  EncryptableDateTime: any;
  EncryptableMarkdown: any;
  EncryptableString: any;
  EncryptableTxHash: any;
  EncryptableURI: any;
  EncryptedPath: any;
  Ens: any;
  EvmAddress: any;
  Handle: any;
  ImageSizeTransform: any;
  Jwt: any;
  Locale: any;
  Markdown: any;
  MimeType: any;
  MomokaId: any;
  MomokaProof: any;
  NftGalleryId: any;
  NftGalleryName: any;
  Nonce: any;
  OnchainPublicationId: any;
  PoapEventId: any;
  ProfileId: any;
  PublicationId: any;
  Signature: any;
  TokenId: any;
  TxHash: any;
  TxId: any;
  URI: any;
  URL: any;
  UUID: any;
  UnixTimestamp: any;
  Void: any;
};
// ...
export type EnsOnchainIdentity = {
  __typename?: "EnsOnchainIdentity";
  /** The default ens mapped to this address */
  name?: Maybe<Scalars["Ens"]>;
};
// ...
export type FeeFollowModuleSettings = {
  __typename?: "FeeFollowModuleSettings";
  /** The amount info */
  amount: Amount;
  contract: NetworkAddress;
  /** The module recipient address */
  recipient: Scalars["EvmAddress"];
  type: FollowModuleType;
};
// ...
export type FollowModule =
  | FeeFollowModuleSettings
  | RevertFollowModuleSettings
  | UnknownFollowModuleSettings;
// ...
export type HandleGuardianResult = {
  __typename?: "HandleGuardianResult";
  cooldownEndsOn?: Maybe<Scalars["DateTime"]>;
  protected: Scalars["Boolean"];
};

export type HandleInfo = {
  __typename?: "HandleInfo";
  /** The full handle - namespace/localname */
  fullHandle: Scalars["Handle"];
  guardian: HandleGuardianResult;
  /** The handle nft token id */
  id: Scalars["TokenId"];
  /** If null its not linked to anything */
  linkedTo?: Maybe<HandleLinkedTo>;
  /** The localname */
  localName: Scalars["String"];
  /** The namespace */
  namespace: Scalars["String"];
  ownedBy: Scalars["EvmAddress"];
  /** The suggested format to use on UI for ease but you can innovate and slice and dice as you want */
  suggestedFormatted: SuggestedFormattedHandle;
};

export type HandleLinkedTo = {
  __typename?: "HandleLinkedTo";
  /** The contract address it is linked to */
  contract: NetworkAddress;
  /** The nft token id it is linked to (this can be the profile Id) */
  nftTokenId: Scalars["TokenId"];
};
// ...
export type ImageSet = {
  __typename?: "ImageSet";
  optimized?: Maybe<Image>;
  raw: Image;
  transformed?: Maybe<Image>;
};
// ...
export type MetadataAttribute = {
  __typename?: "MetadataAttribute";
  key: Scalars["String"];
  /**
   * The type of the attribute. When:
   * - BOOLEAN: the `value` is `true`|`false`
   * - DATE: the `value` is a valid ISO 8601 date string
   * - NUMBER: the `value` is a valid JS number as string
   * - STRING: the `value` is a string.
   * - JSON: the `value` is a valid JSON serialized as string
   *
   */
  type: MetadataAttributeType;
  /** The value serialized as string. It's consumer responsibility to parse it according to `type`. */
  value: Scalars["String"];
};

export enum MetadataAttributeType {
  Boolean = "BOOLEAN",
  Date = "DATE",
  Json = "JSON",
  Number = "NUMBER",
  String = "STRING",
}
// ...
export type NetworkAddress = {
  __typename?: "NetworkAddress";
  address: Scalars["EvmAddress"];
  chainId: Scalars["ChainId"];
};
// ...
export type OptimisticStatusResult = {
  __typename?: "OptimisticStatusResult";
  isFinalisedOnchain: Scalars["Boolean"];
  value: Scalars["Boolean"];
};
// ...
export type Profile = {
  __typename?: "Profile";
  /** When the profile was created */
  createdAt: Scalars["DateTime"];
  /** The follow module */
  followModule?: Maybe<FollowModule>;
  /** The profile follow nft address */
  followNftAddress?: Maybe<NetworkAddress>;
  guardian?: Maybe<ProfileGuardianResult>;
  /** The profile handle - a profile may not have one */
  handle?: Maybe<HandleInfo>;
  /** The profile id */
  id: Scalars["ProfileId"];
  interests: Array<Scalars["String"]>;
  invitedBy?: Maybe<Profile>;
  /** The number of invites left */
  invitesLeft: Scalars["Int"];
  /** The profile metadata. You can optionally query profile metadata by app id.  */
  metadata?: Maybe<ProfileMetadata>;
  /** The on chain identity */
  onchainIdentity: ProfileOnchainIdentity;
  operations: ProfileOperations;
  /** Who owns the profile */
  ownedBy: NetworkAddress;
  /** If the profile has been recommended by the authenticated user */
  peerToPeerRecommendedByMe: Scalars["Boolean"];
  /** If the profile has got signless enabled */
  signless: Scalars["Boolean"];
  /** If lens API will sponsor this persons for gasless experience, note they can have signless on but sponsor false which means it be rejected */
  sponsor: Scalars["Boolean"];
  stats: ProfileStats;
  txHash: Scalars["TxHash"];
};
// ...
export type ProfileGuardianResult = {
  __typename?: "ProfileGuardianResult";
  cooldownEndsOn?: Maybe<Scalars["DateTime"]>;
  protected: Scalars["Boolean"];
};
// ...
export type ProfileMetadata = {
  __typename?: "ProfileMetadata";
  /** The app that this metadata is displayed on */
  appId?: Maybe<Scalars["AppId"]>;
  /** Profile Custom attributes */
  attributes?: Maybe<Array<MetadataAttribute>>;
  /** The bio for the profile */
  bio?: Maybe<Scalars["Markdown"]>;
  /** The cover picture for the profile */
  coverPicture?: Maybe<ImageSet>;
  /** The display name for the profile */
  displayName?: Maybe<Scalars["String"]>;
  /** The picture for the profile */
  picture?: Maybe<ProfilePicture>;
  /** The raw uri for the which the profile metadata was set as */
  rawURI: Scalars["URI"];
};

export type ProfileMirrorResult = {
  __typename?: "ProfileMirrorResult";
  mirrorId: Scalars["PublicationId"];
  mirroredAt: Scalars["DateTime"];
  profile: Profile;
};

export type ProfileOnchainIdentity = {
  __typename?: "ProfileOnchainIdentity";
  /** The ens information */
  ens?: Maybe<EnsOnchainIdentity>;
  /** The POH status */
  proofOfHumanity: Scalars["Boolean"];
  /** The sybil dot org information */
  sybilDotOrg: SybilDotOrgIdentity;
  /** The worldcoin identity */
  worldcoin: WorldcoinIdentity;
};

export type ProfileOperations = {
  __typename?: "ProfileOperations";
  canBlock: Scalars["Boolean"];
  canFollow: TriStateValue;
  canUnblock: Scalars["Boolean"];
  canUnfollow: Scalars["Boolean"];
  hasBlockedMe: OptimisticStatusResult;
  id: Scalars["ProfileId"];
  isBlockedByMe: OptimisticStatusResult;
  isFollowedByMe: OptimisticStatusResult;
  isFollowingMe: OptimisticStatusResult;
};

export type ProfileOwnershipCondition = {
  __typename?: "ProfileOwnershipCondition";
  profileId: Scalars["ProfileId"];
};

export type ProfilePicture = ImageSet | NftImage;
// ...
export type ProfileStats = {
  __typename?: "ProfileStats";
  comments: Scalars["Int"];
  countOpenActions: Scalars["Int"];
  followers: Scalars["Int"];
  following: Scalars["Int"];
  id: Scalars["ProfileId"];
  /** The profile classifier score of this profile relative to others on Lens. It is a % out of 100. */
  lensClassifierScore?: Maybe<Scalars["Float"]>;
  mirrors: Scalars["Int"];
  posts: Scalars["Int"];
  publications: Scalars["Int"];
  quotes: Scalars["Int"];
  /** How many times a profile has reacted on something */
  reacted: Scalars["Int"];
  /** How many times other profiles have reacted on something this profile did */
  reactions: Scalars["Int"];
};
// ...
export type RevertFollowModuleSettings = {
  __typename?: "RevertFollowModuleSettings";
  contract: NetworkAddress;
  type: FollowModuleType;
};
// ...
export type SuggestedFormattedHandle = {
  __typename?: "SuggestedFormattedHandle";
  /** The full formatted handle - namespace/@localname */
  full: Scalars["String"];
  /** The formatted handle - @localname */
  localName: Scalars["String"];
};

export enum SupportedFiatType {
  Eur = "EUR",
  Gbp = "GBP",
  Usd = "USD",
}

export type SupportedModule = KnownSupportedModule | UnknownSupportedModule;

export type SupportedModulesRequest = {
  cursor?: InputMaybe<Scalars["Cursor"]>;
  includeUnknown?: InputMaybe<Scalars["Boolean"]>;
  limit?: InputMaybe<LimitType>;
  onlyVerified?: InputMaybe<Scalars["Boolean"]>;
};

export type SybilDotOrgIdentity = {
  __typename?: "SybilDotOrgIdentity";
  source?: Maybe<SybilDotOrgIdentitySource>;
  /** The sybil dot org status */
  verified: Scalars["Boolean"];
};
// ...
export enum TriStateValue {
  No = "NO",
  Unknown = "UNKNOWN",
  Yes = "YES",
}
// ...
export type UnknownFollowModuleSettings = {
  __typename?: "UnknownFollowModuleSettings";
  contract: NetworkAddress;
  /**
   * The data used to setup the module which you can decode with your known ABI
   * @deprecated Use initializeResultData instead
   */
  followModuleReturnData?: Maybe<Scalars["BlockchainData"]>;
  /** The data used to setup the module */
  initializeCalldata?: Maybe<Scalars["BlockchainData"]>;
  /** The data returned from the init module */
  initializeResultData?: Maybe<Scalars["BlockchainData"]>;
  /** True if the module can be signedless and use lens manager without a signature */
  signlessApproved: Scalars["Boolean"];
  /** True if the module can be sponsored through gasless so the user does not need to pay for gas */
  sponsoredApproved: Scalars["Boolean"];
  type: FollowModuleType;
  /** True if the module is deemed as safe */
  verified: Scalars["Boolean"];
};
// ...
export type WorldcoinIdentity = {
  __typename?: "WorldcoinIdentity";
  /** If the profile has verified as a user */
  isHuman: Scalars["Boolean"];
};